requirements.txt
cryptography==41.0.7
lime==0.2.0.1
numpy==1.24.3
opencv_contrib_python==4.8.0.74
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==10.0.1
Pillow==10.1.0
pydicom==2.4.4
skimage==0.0
streamlit==1.22.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
